add a new series: formatTree. It maintains the branches in tree. So roothMain & tmain accrods this to construct branch. But now tmain part is not yet completed
