+++ histProduce/src/rooTHMainTkTk.cc
@@ -8,10 +8,13 @@
 
 
 root_TreeHistoMain_TkTk::root_TreeHistoMain_TkTk( TFileDirectory* d ) :
-    root_TreeHistoMain( d, "TkTk" ),
+    root_TreeHistoMain( d, "TkTk" ), formatTree_TkTk( totNumD, totNumI ),
     kaonMass ( 0.493667 ), protonMass ( 0.9382720813 ), pionMass ( 0.13957061 )
 {
     setInputTreeName( "lbSpecificDecay/TkTk" );
+
+    if ( !d ) return;
+    SetNewFormatTree(thisTree());
     RegTree();
     RegHisto();
 }
@@ -19,7 +22,6 @@ void root_TreeHistoMain_TkTk::Process( unsigned int pIdx )
 {
     try
     {
-        // preselect events {{{
         Clear();
         readTree()->GetEntry( pIdx );
 
@@ -33,38 +35,14 @@ void root_TreeHistoMain_TkTk::Process( unsigned int pIdx )
     } catch (...) {}
 }
 
-void root_TreeHistoMain_TkTk::Clear()
-{
-    memset( dataD, 0x00, totNumD*sizeof( double ) );
-    memset( dataI, 0x00, totNumI*sizeof( int ) );
-}
 void root_TreeHistoMain_TkTk::RegTree()
 {
-    // don't reg tree.
-    return;
     if ( NoOutput() ) return;
 
     TTree* t = thisTree();
-    t->Branch( "tktkMass", &dataD[tktkMass], "tktkMass/D" );
-    t->Branch( "tktkFD2d", &dataD[tktkFlightDistance2d], "tktkFD2d/D" );
-    t->Branch( "tktkVtxprob", &dataD[tktkVtxprob], "tktkVtxprob/D" );
-    t->Branch( "tktkPt", &dataD[tktkPt], "tktkPt/D" );
-    t->Branch( "fake_Lam0Mass", &dataD[fake_Lam0Mass], "fake_Lam0Mass/D" );
-    t->Branch( "fake_KstarMass", &dataD[fake_KstarMass], "fake_KstarMass/D" );
-    t->Branch( "fake_KshortMass", &dataD[fake_KshortMass], "fake_KshortMass/D" );
-    t->Branch( "fake_PhiMass", &dataD[fake_PhiMass], "fake_PhiMass/D" );
-    t->Branch( "ptkMom", &dataD[ptkMom], "ptkMom/D" );
-    t->Branch( "ptkDEDX.Harmonic", &dataD[ptkDEDX_Harmonic], "ptkDEDX.Harmonic/D" );
-    t->Branch( "ptkDEDX.pixelHrm", &dataD[ptkDEDX_pixelHrm], "ptkDEDX.pixelHrm/D" );
-    t->Branch( "ptkIPt", &dataD[ptkIPt], "ptkIPt/D" );
-    t->Branch( "ptkIPtErr", &dataD[ptkIPtErr], "ptkIPtErr/D" );
-    t->Branch( "ntkMom", &dataD[ntkMom], "ntkMom/D" );
-    t->Branch( "ntkDEDX.Harmonic", &dataD[ntkDEDX_Harmonic], "ntkDEDX.Harmonic/D" );
-    t->Branch( "ntkDEDX.pixelHrm", &dataD[ntkDEDX_pixelHrm], "ntkDEDX.pixelHrm/D" );
-    t->Branch( "ntkIPt", &dataD[ntkIPt], "ntkIPt/D" );
-    t->Branch( "ntkIPtErr", &dataD[ntkIPtErr], "ntkIPtErr/D" );
-
+    RegFormatTree();
     // t->Branch("newBranch", &dataD[totNumD+newBranch], "newBranch/D" );
+    return;
 }
 void root_TreeHistoMain_TkTk::RegHisto()
 {
@@ -76,24 +54,9 @@ void root_TreeHistoMain_TkTk::RegHisto()
 void root_TreeHistoMain_TkTk::LoadSourceBranch()
 {
     TTree* t = readTree();
-    t->SetBranchAddress( "tktkMass", &readD[tktkMass] );
-    t->SetBranchAddress( "tktkFD2d", &readD[tktkFlightDistance2d] );
-    t->SetBranchAddress( "tktkVtxprob", &readD[tktkVtxprob] );
-    t->SetBranchAddress( "tktkPt", &readD[tktkPt] );
-    t->SetBranchAddress( "fake_Lam0Mass", &readD[fake_Lam0Mass] );
-    t->SetBranchAddress( "fake_KstarMass", &readD[fake_KstarMass] );
-    t->SetBranchAddress( "fake_KshortMass", &readD[fake_KshortMass] );
-    t->SetBranchAddress( "fake_PhiMass", &readD[fake_PhiMass] );
-    t->SetBranchAddress( "ptkMom", &readD[ptkMom] );
-    t->SetBranchAddress( "ptkDEDX.Harmonic", &readD[ptkDEDX_Harmonic] );
-    t->SetBranchAddress( "ptkDEDX.pixelHrm", &readD[ptkDEDX_pixelHrm] );
-    t->SetBranchAddress( "ptkIPt", &readD[ptkIPt] );
-    t->SetBranchAddress( "ptkIPtErr", &readD[ptkIPtErr] );
-    t->SetBranchAddress( "ntkMom", &readD[ntkMom] );
-    t->SetBranchAddress( "ntkDEDX.Harmonic", &readD[ntkDEDX_Harmonic] );
-    t->SetBranchAddress( "ntkDEDX.pixelHrm", &readD[ntkDEDX_pixelHrm] );
-    t->SetBranchAddress( "ntkIPt", &readD[ntkIPt] );
-    t->SetBranchAddress( "ntkIPtErr", &readD[ntkIPtErr] );
+    this->SetOldFormatTree(t);
+    this->LoadFormatSourceBranch();
+    return;
 }
 void root_TreeHistoMain_TkTk::SummaryCalc()
 {
